refactor: extract flagUnlabelledClickableElements as a function
